+++ Sources/CareKitEssentials/Cards/Shared/Extensions/CareKitEssentialSliderLogView+EventViewable.swift
@@ -1,0 +1,25 @@
+//
+//  CareKitEssentialSliderLogView+EventViewable.swift
+//  CareKitEssentials
+//
+//  Created by Corey Baker on 5/5/25.
+//  Copyright © 2025 Network Reconnaissance Lab. All rights reserved.
+//
+
+#if !os(watchOS)
+
+import CareKitStore
+
+extension CareKitEssentialSliderLogView: EventViewable {
+
+	public init?(
+		event: OCKAnyEvent,
+		store: any OCKAnyStoreProtocol
+	) {
+		self.init(
+			event: event
+		)
+	}
+}
+
+#endif
